# packages/contracts/openapi/api-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OnePic API",
    "version": "0.1.0",
    "summary": "OnePic 单图模板库的首期受控生图 API",
    "description": "当前 phase-one API 表面（docs/full-stack-architecture.md §6、ADR 0001 D-1）：\n健康、上传、预审、生成/取消/删除/历史、收藏/集合与工作区导出。模板目录仍由\n静态 JSON/TXT/WebP 提供。成功响应为 `{ data, meta }`，失败响应为 `{ error }`；\n字段命名 camelCase，时间 UTC ISO-8601，ID 使用 UUID/ULID，分页使用 cursor。\n本契约是生成类型与运行时校验的唯一事实来源；实现和 generated client 漂移由 CI 阻断。\n"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/generations": {
      "get": {
        "operationId": "listGenerations",
        "tags": [
          "workspace"
        ],
        "summary": "本人生成历史（cursor 分页）",
        "description": "仅返回本人任务，稳定排序 created_at DESC + id tiebreaker。 cursor 为不透明签名串（服务端 HMAC 签发，篡改/非法一律 400）， 逐页遍历不重不漏；state 过滤可选。limit 上限 50，默认 20。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/HistoryStateFilter"
          },
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "一页历史（meta.nextCursor 存在时还有下一页）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfGenerationList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          }
        }
      },
      "post": {
        "operationId": "createGeneration",
        "tags": [
          "generation"
        ],
        "summary": "幂等创建生成任务",
        "description": "携带 Idempotency-Key；同键同请求指纹返回同一任务，同键不同指纹返回 409 IDEMPOTENCY_CONFLICT。服务端重新读取不可变模板版本并比对 promptSha256，拒绝浏览器正文替换。任务异步执行，恒返回 202， 经 GET /generations/{generationId} 轮询。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "幂等键；浏览器端在一次点击周期内生成并复用。",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "任务已受理（新建或幂等重放）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfGenerationStatus"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "409": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "429": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
...
    "/collections": {
      "get": {
        "operationId": "listCollections",
        "tags": [
          "workspace"
        ],
        "summary": "本人集合列表（cursor 分页，含条目计数）",
        "description": "仅返回本人集合，稳定排序 created_at ASC + id tiebreaker； cursor 机制与历史分页一致（签名防篡改，非法 400）。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "一页集合",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfCollectionList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          }
        }
      },
      "post": {
        "operationId": "createCollection",
        "tags": [
          "workspace"
        ],
        "summary": "创建集合",
        "description": "同一属主下 name 唯一（§1.13）；重名返回 409 COLLECTION_NAME_CONFLICT。",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "集合已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfCollectionSummary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "409": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/collections/{collectionId}": {
      "delete": {
        "operationId": "deleteCollection",
        "tags": [
          "workspace"
        ],
        "summary": "删除集合（不触碰任务与媒体）",
        "description": "仅删除集合行与成员关系（collection_item 级联）；绝不删除 generation / media_object / result 行，被收藏任务仍可经 GET /generations/{id} 读取。仅属主可删（越权 403，不存在 404）。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CollectionId"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "集合已删除",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfCollectionDeleted"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/collections/{collectionId}/items": {
      "post": {
        "operationId": "addCollectionItem",
        "tags": [
          "workspace"
        ],
        "summary": "收藏模板或生成任务到集合（幂等）",
        "description": "PK (collection_id, item_type, item_key) 保证重复收藏幂等：重复 POST 返回相同结果（200，addedAt 为首次收藏时间），不产生重复行。 itemType=template 时 itemKey 必须存在于已导入目录；itemType= generation 时 itemKey 必须是本人 generation 的 UUID（他人 403， 不存在 404）。集合本身仅属主可写（越权 403，不存在 404）。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CollectionId"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionItemAddRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "条目已在集合中（新增或幂等重放）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfCollectionItem"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/collections/{collectionId}/items/{itemType}/{itemKey}": {
      "delete": {
        "operationId": "removeCollectionItem",
        "tags": [
          "workspace"
        ],
        "summary": "从集合移除条目（幂等）",
        "description": "条目不存在同样返回 200（removed=false）；仅属主可操作。",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CollectionId"
          },
          {
            "$ref": "#/components/parameters/CollectionItemType"
          },
          {
            "$ref": "#/components/parameters/CollectionItemKey"
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "移除结果（removed 表示是否确有此行）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfCollectionItemRemoval"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/ApiFailureResponse"
          },
          "404": {
            "$ref": "#/components/responses/ApiFailureResponse"
          }
        }
      }
    },
    "/exports/workspace": {
      "get": {
        "operationId": "exportWorkspace",
        "tags": [
          "workspace"
        ],
        "summary": "导出本人收藏/集合/历史（不含密钥）",
        "description": "data 严格遵循 docs/design/backend-schemas/workspace-export.schema.json： favorites 为本人集合内收藏的模板 id（去重），collections 为集合名与 成员，history 为本人全部任务摘要（服务端内部以 W03 keyset 分页翻完 全部页）。响应绝不包含密钥、会话、签名材料或他人记录。\n",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "导出文档",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessOfWorkspaceExport"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "onepic_session",
        "description": "服务端 opaque 会话（OIDC 授权码 + PKCE 登录后签发，B03）。"
      }
    },
    "parameters": {
      "CsrfHeader": {
        "name": "x-onepic-requested-with",
        "in": "header",
        "required": true,
        "description": "CSRF 护栏自定义头，恒为 onepic-fetch（跨站表单无法设置）。",
        "schema": {
          "type": "string",
          "const": "onepic-fetch"
        }
      },
...
      "CollectionId": {
        "name": "collectionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "CollectionItemType": {
        "name": "itemType",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "enum": [
            "template",
            "generation"
          ]
        }
      },
      "CollectionItemKey": {
        "name": "itemKey",
        "in": "path",
        "required": true,
        "description": "template 为模板 ID，generation 为任务 UUID。",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      },
      "PageCursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "上一页 meta.nextCursor 原样回传；不透明签名串，篡改/非法一律 400。",
        "schema": {
          "type": "string",
          "maxLength": 512
        }
      },
      "PageLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "每页条数，默认 20，上限 50。",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 50
        }
      },
      "HistoryStateFilter": {
        "name": "state",
        "in": "query",
        "required": false,
        "description": "按任务状态过滤（架构 §9 生命周期枚举）。",
        "schema": {
          "type": "string",
          "enum": [
            "created",
            "queued",
            "running",
            "succeeded",
            "failed",
            "cancelled",
            "expired",
            "outcome_unknown"
          ]
        }
      }
    },
    "responses": {
      "ValidationFailed": {
        "description": "请求未通过契约校验（含未知字段）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiFailure"
            }
          }
        }
      },
      "Unauthenticated": {
        "description": "无有效会话",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiFailure"
            }
          }
        }
      },
      "ApiFailureResponse": {
        "description": "稳定错误码信封（错误码目录见 D06 数据字典）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiFailure"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiMeta": {
        "type": "object",
        "description": "与 data 并列的分页/追踪元数据；结构随端点补充。",
        "additionalProperties": true
      },
...
      "ApiFailure": {
        "type": "object",
        "description": "失败响应信封。",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        }
      },
...
      "GenerationCreateRequest": {
        "type": "object",
        "description": "字段以 docs/design/backend-schemas/generation-create-request.schema.json 为准。",
        "additionalProperties": false,
        "required": [
          "templateId",
          "templateVersion",
          "promptSha256",
          "sourceObjectId",
          "settings"
        ],
        "properties": {
          "templateId": {
            "type": "string",
            "pattern": "^(case-[0-9]+|framework-[0-9]{3})$"
          },
          "templateVersion": {
            "type": "integer",
            "minimum": 1
          },
          "promptSha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "description": "客户端核验过的目录 promptSha256；服务端与不可变版本再比对。"
          },
          "sourceObjectId": {
            "type": "string",
            "format": "uuid"
          },
          "precheckId": {
            "type": "string",
            "format": "uuid"
          },
          "settings": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "model"
            ],
            "properties": {
              "model": {
                "type": "string",
                "minLength": 1
              },
              "quality": {
                "type": "string",
                "enum": [
                  "standard",
                  "high"
                ]
              },
              "aspectRatio": {
                "type": "string",
                "pattern": "^[0-9]+:[0-9]+$"
              }
            }
          }
        }
      },
...
      "GenerationStatus": {
        "type": "object",
        "description": "data 字段以 generation-status-response.schema.json 为准。",
        "additionalProperties": false,
        "required": [
          "id",
          "state",
          "templateId",
          "templateVersion",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "created",
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled",
              "expired",
              "outcome_unknown"
            ]
          },
          "templateId": {
            "type": "string",
            "pattern": "^(case-[0-9]+|framework-[0-9]{3})$"
          },
          "templateVersion": {
            "type": "integer",
            "minimum": 1
          },
          "errorCode": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "result": {
            "$ref": "#/components/schemas/GenerationResult"
          }
        }
      },
      "GenerationStatusMeta": {
        "type": "object",
        "description": "轮询建议与签名下载地址（结果媒体存活时才有 downloadUrl）。",
        "additionalProperties": true,
        "properties": {
          "pollAfterMs": {
            "type": "integer",
            "minimum": 250
          },
          "downloadUrl": {
            "type": "string",
            "description": "短期签名下载路径（/api/v1/media/...，属主绑定）。"
          },
          "downloadExpires": {
            "type": "integer",
            "description": "downloadUrl 的过期时间（Unix 秒）。"
          }
        }
      },
...
      "ApiSuccessOfGenerationStatus": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GenerationStatus"
          },
          "meta": {
            "$ref": "#/components/schemas/GenerationStatusMeta"
          }
        }
      },
...
      "GenerationList": {
        "type": "object",
        "description": "一页历史；item 形状与 generation-status-response 相同（不含 result）。",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenerationStatus"
            }
          }
        }
      },
      "GenerationListMeta": {
        "type": "object",
        "description": "分页元数据；nextCursor 仅在还有下一页时出现。",
        "additionalProperties": true,
        "properties": {
          "nextCursor": {
            "type": "string"
          }
        }
      },
      "ApiSuccessOfGenerationList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GenerationList"
          },
          "meta": {
            "$ref": "#/components/schemas/GenerationListMeta"
          }
        }
      },
      "CollectionSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "createdAt",
          "itemCount"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "itemCount": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "CollectionList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionSummary"
            }
          }
        }
      },
      "ApiSuccessOfCollectionList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CollectionList"
          },
          "meta": {
            "$ref": "#/components/schemas/GenerationListMeta"
          }
        }
      },
      "CollectionCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          }
        }
      },
      "ApiSuccessOfCollectionSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CollectionSummary"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "CollectionDeleted": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "deleted"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean",
            "const": true
          }
        }
      },
      "ApiSuccessOfCollectionDeleted": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CollectionDeleted"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "CollectionItemAddRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "itemType",
          "itemKey"
        ],
        "properties": {
          "itemType": {
            "type": "string",
            "enum": [
              "template",
              "generation"
            ]
          },
          "itemKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "description": "template 为模板 ID（case-N / framework-NNN），generation 为任务 UUID。"
          }
        }
      },
      "CollectionItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "collectionId",
          "itemType",
          "itemKey",
          "addedAt"
        ],
        "properties": {
          "collectionId": {
            "type": "string",
            "format": "uuid"
          },
          "itemType": {
            "type": "string",
            "enum": [
              "template",
              "generation"
            ]
          },
          "itemKey": {
            "type": "string"
          },
          "addedAt": {
            "type": "string",
            "format": "date-time",
            "description": "首次收藏时间；幂等重放不改变该值。"
          }
        }
      },
      "ApiSuccessOfCollectionItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CollectionItem"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "CollectionItemRemoval": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "collectionId",
          "itemType",
          "itemKey",
          "removed"
        ],
        "properties": {
          "collectionId": {
            "type": "string",
            "format": "uuid"
          },
          "itemType": {
            "type": "string",
            "enum": [
              "template",
              "generation"
            ]
          },
          "itemKey": {
            "type": "string"
          },
          "removed": {
            "type": "boolean",
            "description": "是否确有此行被删除；重复移除为 false（幂等 200）。"
          }
        }
      },
      "ApiSuccessOfCollectionItemRemoval": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CollectionItemRemoval"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
...
      "ApiSuccessOfWorkspaceExport": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WorkspaceExport"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      }
    }
  }
}